Testing the sensitivity of the algorithm
Lots of cleanup in this commit. I've been playing with bootstrapping over many random gene searches to try to make this more sensitive. In doing so, I noticed that those conditions with low p-values in the artificial data are just the ones with more genes dedicated to them (makes sense...). Need some way to penalize bad ranks more, assess how surprising the given p-value is or etc.